requirements.txt
fastapi==0.111.*
uvicorn[standard]==0.29.*
python-telegram-bot==21.*
apscheduler==3.10.*
yfinance==0.2.*
httpx==0.27.*
jinja2==3.1.*
pytz==2024.*
